* origin/master: (692 commits)
  textboxes didn't update correctly in some cases
  changing alpha means we need to mark picture dirty
  mark slideshow window initially dirty
  fixed: platforms previously building without sdl and with HAS_XBMC_MUTEX were broken by 846506d9.
  no need to SetTransform/RenderResolution in dim screensaver now that we correctly handle it in the base class
  ensure we run DoRender at the window level - else animations aren't correctly rendered in some circumstance.  Fixes #11660
  ensure we set the window manager as init'd before we load windows not otherwise loaded on demand, ensuring their windwo <visible> conditions are defined. Fixes #11692
  no need for dim screensaver and debug info windows to not set load on demand (there's nothing to load)
  process fadelabel control so that it's not marked dirty when not scrolling
  compute correct render rect for the label control - fixes #11651 (labels without <height> specified)
  ensure we invalidate the grouplist and mark dirty if scrolling.  fixes #11688
  mouse click in an empty portion of the panel container would cause the focused item to be selected
  proper marking of dirty regions for textbox
  cleanup GUITextBox processing so that everything happens in Process
  ensure we use a clean state for fade out transform in textbox - fixes #11670
  fix #11691: LinuxRendererGLES.cpp: adapt to new location of kai toast dialog member, closes #11691
  fixed: crash on picture load due to invalid WaitForSingleObject change in 8c7bdfa
  slight optimisation of AddTransform
  change volume control behaviour
  flag transforms that are the identity as such and skip any multiplication
  ...

Conflicts:
	configure.in
	lib/DllWAVPack.h
	xbmc/cores/VideoRenderers/LinuxRendererGLES.h
	xbmc/cores/dvdplayer/DVDSubtitles/DllLibass.h
	xbmc/input/KeyboardStat.cpp
	xbmc/threads/Atomics.cpp
	xbmc/threads/XBMC_mutex.cpp
	xbmc/windowing/WinEvents.h
	xbmc/xbmc.cpp

+++ lib/ffmpeg/build_xbmc_win32.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-if [ -d .libs ]
+if [ "$1" == "clean" ]
 then
-rm -r .libs
+  if [ -d .libs ]
+  then
+    rm -r .libs
+  fi
+  make distclean
 fi
 
-if [ -f config.mak ]
-then
-make distclean
+if [ ! -d .libs ]; then
+  mkdir .libs
 fi
 
 OPTIONS="
@@ -32,18 +35,18 @@ OPTIONS="
 --enable-encoder=ac3 \
 --enable-encoder=aac \
 --enable-runtime-cpudetect \
+--enable-avfilter \
 --disable-debug \
 --disable-doc"
 
 ./configure --extra-cflags="-fno-common -Iinclude-xbmc-win32/dxva2" --extra-ldflags="-L/xbmc/system/players/dvdplayer" ${OPTIONS} &&
  
-make && 
-mkdir .libs &&
+make &&
 cp lib*/*.dll .libs/ &&
-mv .libs/swscale-0.dll .libs/swscale-0.6.1.dll &&
 cp .libs/avcodec-52.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avcore-0.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avformat-52.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avutil-50.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/avfilter-1.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/postproc-51.dll /xbmc/system/players/dvdplayer/ &&
-cp .libs/swscale-0.6.1.dll /xbmc/system/players/dvdplayer/
+cp .libs/swscale-0.dll /xbmc/system/players/dvdplayer/
